Add ability to scan by service UUID on iOS + macOS
The API has been modified to add a new `forServices:` named parameter to
the `QuickBlue.scan(..)` method. Pass in a list of primary service UUIDs
to only scan for devices advertising one of thoese services. Without
this, there was no way to identify a device specific to your app.

NOTE: The new parameter is only effective on Apple platforms. On Linux
it is currently ignored and on Android and Windows it will probably
crash as I've not modified the platform channel code at all.